Conflicts:
	source/SAMRAI/hier/BaseGridGeometry.C
	source/SAMRAI/hier/Box.C
	source/SAMRAI/hier/Box.h
	source/SAMRAI/hier/BoxUtilities.C
	source/SAMRAI/hier/Index.h
	source/SAMRAI/hier/IntVector.h
	source/SAMRAI/mesh/BalanceBoxBreaker.C
	source/SAMRAI/mesh/BergerRigoutsosNode.C
	source/SAMRAI/mesh/BoxTransitSet.h
	source/SAMRAI/pdat/FaceComplexConstantRefine.C
	source/SAMRAI/pdat/FaceDoubleConstantRefine.C
	source/SAMRAI/pdat/FaceFloatConstantRefine.C
	source/SAMRAI/pdat/FaceIntegerConstantRefine.C
	source/SAMRAI/pdat/OuterfaceComplexConstantRefine.C
	source/SAMRAI/pdat/OuterfaceDoubleConstantRefine.C
	source/SAMRAI/pdat/OuterfaceFloatConstantRefine.C
	source/SAMRAI/pdat/OuterfaceIntegerConstantRefine.C
	source/test/applications/ConvDiff/test_inputs/test.3d.boxes.0000001.0000000
	source/test/applications/ConvDiff/test_inputs/test.3d.boxes.0000002.0000000
	source/test/applications/ConvDiff/test_inputs/test.3d.boxes.0000002.0000001
	source/test/applications/ConvDiff/test_inputs/test.3d.boxes.0000004.0000000
	source/test/applications/ConvDiff/test_inputs/test.3d.boxes.0000004.0000001
	source/test/applications/ConvDiff/test_inputs/test.3d.boxes.0000004.0000002
	source/test/applications/ConvDiff/test_inputs/test.3d.boxes.0000004.0000003

+++ source/SAMRAI/mesh/CascadePartitionerTree.C
@@ -668,7 +668,7 @@ CascadePartitionerTree::recomputeLeafData()
 void
 CascadePartitionerTree::resetObligation(double avg_load)
 {
-   d_obligation = avg_load * size();
+   d_obligation = avg_load * static_cast<double>(size());
    d_group_may_supply = estimatedSurplus() > d_common->d_pparams->getLoadComparisonTol();
 
    if (d_children[0]) {
@@ -692,9 +692,9 @@ CascadePartitionerTree::printClassData(std::ostream& co, const std::string& bord
       << "  near=" << d_near << "  far=" << d_far
       << '\n' << indent
       << "contact=" << d_contact[0] << ',' << d_contact[1]
-      << "  work=" << d_work << '/' << d_obligation << " (" << size() * d_common->d_global_load_avg
+      << "  work=" << d_work << '/' << d_obligation << " (" << static_cast<double>(size()) * d_common->d_global_load_avg
       << ")  estimated surplus=" << estimatedSurplus()
-      << " (" << d_work - (size() * d_common->d_global_load_avg)
+      << " (" << d_work - (static_cast<double>(size()) * d_common->d_global_load_avg)
       << ")  local_load=" << d_common->d_local_load->getSumLoad()
       << '\n' << indent
       << "group_may_supply=" << d_group_may_supply
